Initial version of tables under Git
These files are the first versions of the CMOR tables released to
the CCMI groups on April 28th, 2014. Note that the filenames of
the original set all had the ending '_comp-v6.txt'.  The only
change here is to delete that bit of text.

Also included are the comma-separated-value files exported from
the Excel spreadsheet version of the data request used to
construct the tables.