August 6, 2014 - corrections to the original release.
A large number of (hopefully minor) changes collected since
the first version of the CMOR tables were released to the CCMI
community on April 28, 2014.  Thanks to Martin Juckes (BADC),
Jamie Kettleborough (UKMet) and Steve Hardiman (UKMet) for
their input in developing these.

General changes

The table names were shortened to be of the form expected by
CMOR, in conjunction with moving version control of the tables
to Git.

Rolled back the CF version specified in the header of all tables
to specify 1.4, because CMOR issues a long list of warnings when
the table version does not match the one specified internally.

Removed the specification of cell_measures: area: areacella for
all zonal average fields (monthly and daily tables), the daily
PSC area fields and the annual 0-D chemical budget fields because
they are not on a lat-lon grid for which the areacella field is
valid.

Added an alternate time axis - there is now time and time1. The
time axis must have bounds set, while the time1 axis does not
enforce bounds.  The addition of time1 allows for a simplified
writing of instantaneous fields.  All fields specified as an
average over a particular time period (monthly, annual) have not
been changed and must still have bounds supplied for the time axis.
For the daily and hourly fields the time axis is set to time1 for
which bounds are not required. In the absence of bounds on time1
the fields are assumed instantaneous.  If the fields being
submitted are averages over a particular time period, for example
24-hour averages for the daily fields, bounds should be set
nonetheless.  Note that the daily 3-D fields specifically request
instantaneous fields and this has been enforced by settings
'cell_methods: time: point' along with the use of the time1 axis.
The one exception are fields requesting daily maximum or minimum
values (e.g. tasmin) which must have bounds set for the time axis
and still use the time axis.

Addition of cell_methods: area: sum to fields that represent
integrals over area, such as the PSC area fields and the annual
global chemical budget terms.

Specific changes:

For the vgd variable, vegetation_area_fraction, in the monthly
table, the vegetation axis was moved to be before the time axis
in the dimensions list since it is not a scalar dimension.

The sfta variable, surface_temperature, in the daily table was
renamed ts to harmonize with a similar variable in CMIP5. The
comment for this variable was changed as well.