CCMI1 table updates to October 31, 2014
Mostly minor changes to long_name and comment attributes and a few
instances where quantities had units specified as years have been
changed to year, for consistency. These changes are not expected to
have any impact on the users of the tables or the data produced by
the table.  There are also a couple of important changes, which are:

1)  the units specified for cloud_area_fraction (clt) and
convective_cloud_area_fraction (convclt) in the CCMI1_monthly table
have been changed from a unitless fraction (1) to percentage (%) to
be consistent with the CCMI1_daily table and CMIP5 tables

2) the upward_eastward_momentum_flux_in_air_due_to_nonorographic
variables for eastward and westward gravity waves (nogwfluxw and
nogwfluxe) were switched because it looks like there were errors in
the CCMVal-2 request where these were originally specified.  Comments
on these have also been changed. The table now has:

    nogwfluxw = ...westward_gravity_waves
    nogwfluxe = ...eastward_gravity_waves